+++ scripts/__init__.py
@@ -1,5 +1,6 @@
 from .extractAllEditions import main as editionExtraction
 from .updateOldInTechOpenLinks import main as updateInTechOpenLinks
-from .updateESEditionFormats import main as updateESEditionFormats
+from .updateESEditionFormats import main as updateESEditionFormat
 from .updateLinksCatalog import main as updateLinks
 from .updateRecordsCatalogLink import main as updateRecordCatalogs
+from .govDocES import main as govDocUpdate