+++ CHANGELOG.md
@@ -6,9 +6,12 @@
 - Exposed entire authors object to the edition details page
 - Two new scripts to update NYPL catalogs links in Record and Link table
 - Updated NYPL mapping and formats in APIUtils with new catalog link
+- Government document filter for advanced search page
+- Script to update past works in ES with the proper boolean value in their is_government_document field
 ### Fixed
 - Added improved error handling around unexpected `media_type` values
 - Improve handling of NYPL Catalog records without EDD links
+- Future government document works in ES now have 'True' boolean value in their is_government_document field
 
 ## 2022-06-09 -- v0.10.4
 ### Added